Change type

+++ rs/backend/src/stats.rs
@@ -122,7 +122,7 @@ pub fn encode_metrics(w: &mut MetricsEncoder<Vec<u8>>) -> std::io::Result<()> {
 pub fn stable_memory_size_bytes() -> u64 {
     #[cfg(target_arch = "wasm32")]
     {
-        stable64_size() * WASM_PAGE_SIZE_IN_BYTES
+        stable64_size() * (WASM_PAGE_SIZE_IN_BYTES as u64)
     }
     #[cfg(not(target_arch = "wasm32"))]
     {
@@ -134,7 +134,7 @@ pub fn stable_memory_size_bytes() -> u64 {
 pub fn wasm_memory_size_bytes() -> u64 {
     #[cfg(target_arch = "wasm32")]
     {
-        (wasm_memory_size(0) as u64) * WASM_PAGE_SIZE_IN_BYTES
+        (wasm_memory_size(0) as u64) * (WASM_PAGE_SIZE_IN_BYTES as u64)
     }
     // This can happen only for test builds.  When compiled for a canister, the target is
     // always wasm32.
